Move enable notification logic into a service file
